On the all tab have a show X button
This was needed so the navigation is clear.  There are some small
problems with the back button not working if you click on say an event
and then try and go back to the updates page.  I'll have to look at this
later.  I don't know why the browser won't go back if there is an anchor
in the url...